requirements.txt
pyqt5
esptool
pyserial
appdirs
